+++ examples/data-sources/conjur_secret/data-source.tf
@@ -1,3 +1,3 @@
-data "conjur_secret" "cloud_dbpass" {
+data "conjur_secret" "my_secret" {
   name = data/test/var-name
 }