+++ internal/provider/resource_authenticator_test.go
@@ -1,15 +1,36 @@
 package provider
 
 import (
+	"context"
 	"encoding/json"
 	"testing"
 
 	"github.com/cyberark/conjur-api-go/conjurapi"
 	"github.com/hashicorp/terraform-plugin-framework/attr"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+func TestConjurAuthenticatorResource_Schema(t *testing.T) {
+	t.Parallel()
+
+	ctx := context.Background()
+	ds := NewConjurAuthenticatorResource()
+
+	schemaRequest := resource.SchemaRequest{}
+	schemaResponse := &resource.SchemaResponse{}
+
+	ds.Schema(ctx, schemaRequest, schemaResponse)
+	if schemaResponse.Diagnostics.HasError() {
+		t.Fatalf("Schema diagnostics had errors: %+v", schemaResponse.Diagnostics)
+	}
+
+	if diagnostics := schemaResponse.Schema.ValidateImplementation(ctx); diagnostics.HasError() {
+		t.Fatalf("Schema validation failed: %+v", diagnostics)
+	}
+}
 
 func TestConjurAuthenticatorResource_buildAuthenticatorPayload(t *testing.T) {
 	r := &ConjurAuthenticatorResource{}