+++ internal/provider/resource_host_test.go
@@ -5,11 +5,31 @@ import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-framework/attr"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+func TestConjurHostResource_Schema(t *testing.T) {
+	t.Parallel()
+
+	ctx := context.Background()
+	ds := NewConjurHostResource()
+
+	schemaRequest := resource.SchemaRequest{}
+	schemaResponse := &resource.SchemaResponse{}
+
+	ds.Schema(ctx, schemaRequest, schemaResponse)
+	if schemaResponse.Diagnostics.HasError() {
+		t.Fatalf("Schema diagnostics had errors: %+v", schemaResponse.Diagnostics)
+	}
+
+	if diagnostics := schemaResponse.Schema.ValidateImplementation(ctx); diagnostics.HasError() {
+		t.Fatalf("Schema validation failed: %+v", diagnostics)
+	}
+}
 
 func TestConjurHostResource_buildHostPayload(t *testing.T) {
 	r := &ConjurHostResource{}